Split the linter into several files (private)
