+++ lib/linter/jdl_linter.js
@@ -16,16 +16,12 @@
  * limitations under the License.
  */
 
-const _ = require('lodash');
 const { readFile } = require('../readers/file_reader');
 const JDLReader = require('../readers/jdl_reader');
-const Rules = require('./rules');
 const Issues = require('./issues/issues');
-const EntityIssue = require('./issues/entity_issue');
-const FieldIssue = require('./issues/field_issue');
-const EnumIssue = require('./issues/enum_issue');
-const RelationshipIssue = require('./issues/relationship_issue');
-const RelationshipTypes = require('../core/jhipster/relationship_types');
+const { checkEntities } = require('./entity_linter');
+const { checkEnums } = require('./enum_linter');
+const { checkRelationships } = require('./relationship_linter');
 
 module.exports = {
   createJDLLinterFromContent,
@@ -76,199 +72,14 @@ function makeJDLLinter(content) {
   };
 }
 
-// entities
-
 function checkForEntityDeclarationIssues() {
-  if (!cst.children.entityDeclaration) {
-    return;
-  }
-  checkForDuplicatedEntities(cst.children.entityDeclaration);
-  cst.children.entityDeclaration.forEach(entityDeclaration => {
-    checkForUselessEntityBraces(entityDeclaration);
-    checkForUselessTableName(entityDeclaration);
-    const entityName = entityDeclaration.children.NAME[0].image;
-    const fieldDeclarations = getFieldDeclarationsFromEntity(entityDeclaration);
-    if (fieldDeclarations.length !== 0) {
-      checkForDuplicatedFields(entityName, fieldDeclarations);
-    }
-  });
+  checkEntities(cst, issues);
 }
-
-function checkForDuplicatedEntities(entityDeclarations) {
-  const entityNames = new Set();
-  const duplicatedEntityIssues = new Map(); // key: entityName, value: issue
-  entityDeclarations.forEach(entityDeclaration => {
-    const entityName = entityDeclaration.children.NAME[0].image;
-    if (entityNames.has(entityName)) {
-      if (!duplicatedEntityIssues.has(entityName)) {
-        duplicatedEntityIssues.set(
-          entityName,
-          new EntityIssue({
-            ruleName: Rules.RuleNames.ENT_DUPLICATED,
-            entityName
-          })
-        );
-      }
-    } else {
-      entityNames.add(entityName);
-    }
-  });
-  duplicatedEntityIssues.forEach(issue => {
-    issues.addEntityIssue(issue);
-  });
-}
-
-function checkForUselessEntityBraces(entityDeclaration) {
-  const entityBody = entityDeclaration.children.entityBody;
-  if (entityBody && Object.keys(entityBody[0].children).length === 2) {
-    // only LCURLY & RCURLY
-    issues.addEntityIssue(
-      new EntityIssue({
-        ruleName: Rules.RuleNames.ENT_SHORTER_DECL,
-        entityName: entityDeclaration.children.NAME[0].image
-      })
-    );
-  }
-}
-
-function checkForUselessTableName(entityDeclaration) {
-  const entityName = entityDeclaration.children.NAME[0].image;
-  const entityTableNameDeclaration = entityDeclaration.children.entityTableNameDeclaration;
-  if (entityTableNameDeclaration) {
-    const tableName = entityTableNameDeclaration[0].children.NAME[0].image;
-    if (_.snakeCase(entityName) === tableName) {
-      issues.addEntityIssue(
-        new EntityIssue({
-          ruleName: Rules.RuleNames.ENT_OPTIONAL_TABLE_NAME,
-          entityName
-        })
-      );
-    }
-  }
-}
-
-function getFieldDeclarationsFromEntity(entityDeclaration) {
-  if (entityDeclaration.children.entityBody && entityDeclaration.children.entityBody[0].children.fieldDeclaration) {
-    return entityDeclaration.children.entityBody[0].children.fieldDeclaration;
-  }
-  return [];
-}
-
-function checkForDuplicatedFields(entityName, fieldDeclarations) {
-  const fieldNames = new Set();
-  const duplicatedFieldIssues = new Map(); // key: fieldName, value: issue
-  fieldDeclarations.forEach(fieldDeclaration => {
-    const fieldName = fieldDeclaration.children.NAME[0].image;
-    if (fieldNames.has(fieldName)) {
-      if (!duplicatedFieldIssues.has(fieldName)) {
-        duplicatedFieldIssues.set(
-          fieldName,
-          new FieldIssue({
-            ruleName: Rules.RuleNames.FLD_DUPLICATED,
-            fieldName,
-            entityName
-          })
-        );
-      }
-    } else {
-      fieldNames.add(fieldName);
-    }
-  });
-  duplicatedFieldIssues.forEach(issue => {
-    issues.addFieldIssue(issue);
-  });
-}
-
-// enums
 
 function checkForEnumDeclarationIssues() {
-  if (!cst.children.enumDeclaration) {
-    return;
-  }
-  checkForDuplicatedEnums(cst.children.enumDeclaration);
-  if (cst.children.entityDeclaration) {
-    const fieldDeclarations = getAllFieldDeclarations();
-    checkForUnusedEnums(cst.children.enumDeclaration, fieldDeclarations);
-  }
+  checkEnums(cst, issues);
 }
-
-function checkForDuplicatedEnums(enumDeclarations) {
-  const enumNames = new Set();
-  const duplicatedEnumIssues = new Map(); // key: enumName, value: issue
-  enumDeclarations.forEach(enumDeclaration => {
-    const enumName = enumDeclaration.children.NAME[0].image;
-    if (enumNames.has(enumName)) {
-      if (!duplicatedEnumIssues.has(enumName)) {
-        duplicatedEnumIssues.set(
-          enumName,
-          new EnumIssue({
-            ruleName: Rules.RuleNames.ENUM_DUPLICATED,
-            enumName
-          })
-        );
-      }
-    } else {
-      enumNames.add(enumName);
-    }
-  });
-  duplicatedEnumIssues.forEach(issue => {
-    issues.addEnumIssue(issue);
-  });
-}
-
-function getAllFieldDeclarations() {
-  return cst.children.entityDeclaration.reduce((fieldDeclarations, entityDeclaration) => {
-    return fieldDeclarations.concat(getFieldDeclarationsFromEntity(entityDeclaration));
-  }, []);
-}
-
-function checkForUnusedEnums(enumDeclarations, fieldDeclarations) {
-  const usedEnums = fieldDeclarations.map(fieldDeclaration => fieldDeclaration.children.type[0].children.NAME[0].image);
-  const declaredEnums = new Set(enumDeclarations.map(enumDeclaration => enumDeclaration.children.NAME[0].image));
-  usedEnums.forEach(usedEnum => {
-    declaredEnums.delete(usedEnum);
-  });
-  if (declaredEnums.size !== 0) {
-    declaredEnums.forEach(unusedEnum => {
-      issues.addEnumIssue(new EnumIssue({ enumName: unusedEnum, ruleName: Rules.RuleNames.ENUM_UNUSED }));
-    });
-  }
-}
-
-// relationships
 
 function checkForRelationshipIssues() {
-  if (!cst.children.relationDeclaration || cst.children.relationDeclaration.length === 0) {
-    return;
-  }
-  checkForCollapsibleRelationships(cst.children.relationDeclaration);
-}
-
-function checkForCollapsibleRelationships(relationshipDeclarations) {
-  const sortedRelationships = {
-    [RelationshipTypes.ONE_TO_ONE]: [],
-    [RelationshipTypes.ONE_TO_MANY]: [],
-    [RelationshipTypes.MANY_TO_ONE]: [],
-    [RelationshipTypes.MANY_TO_MANY]: []
-  };
-  relationshipDeclarations.forEach(relationshipDeclaration => {
-    const type = RelationshipTypes[Object.keys(relationshipDeclaration.children.relationshipType[0].children)[0]];
-    const from = relationshipDeclaration.children.relationshipBody[0].children.from[0].children.NAME[0].image;
-    const to = relationshipDeclaration.children.relationshipBody[0].children.to[0].children.NAME[0].image;
-    sortedRelationships[type].push({ from, to });
-  });
-  Object.keys(sortedRelationships).forEach(type => {
-    if (sortedRelationships[type].length > 1) {
-      sortedRelationships[type].forEach(relationship => {
-        issues.addRelationshipIssue(
-          new RelationshipIssue({
-            ruleName: Rules.RuleNames.REL_INDIVIDUAL_DECL,
-            from: relationship.from,
-            to: relationship.to,
-            type
-          })
-        );
-      });
-    }
-  });
+  checkRelationships(cst, issues);
 }
